+++ src/index.ts
@@ -1,7 +1,21 @@
-export * from './types.js';
-export { parsePlaylist } from './parser.js';
-export { normalizeEntry, normalizePlaylist } from './normalize.js';
-export { isXtreamUrl, parseXtream, makeXtreamCredentials, buildXtreamM3uUrl, buildXtreamCatchupUrl } from './xtream.js';
-export { fetchText, loadPlaylistFromUrl } from './http.js';
-export { parseXmltv, parseXmltvPrograms, buildEpgBindingIndex, buildChannelCategoryMap, parseXmltvDate, type TvgChannel, type TvgProgramme } from './xmltv.js';
-export { enrichPlaylistWithEpg, type EnrichOptions } from './enrich.js';
+export * from "./types.js";
+export { parsePlaylist } from "./parser.js";
+export { normalizeEntry, normalizePlaylist } from "./normalize.js";
+export {
+  isXtreamUrl,
+  parseXtream,
+  makeXtreamCredentials,
+  buildXtreamM3uUrl,
+  buildXtreamCatchupUrl,
+} from "./xtream.js";
+export { fetchText, loadPlaylistFromUrl } from "./http.js";
+export {
+  parseXmltv,
+  parseXmltvPrograms,
+  buildEpgBindingIndex,
+  buildChannelCategoryMap,
+  parseXmltvDate,
+  type TvgChannel,
+  type TvgProgramme,
+} from "./xmltv.js";
+export { enrichPlaylistWithEpg, type EnrichOptions } from "./enrich.js";